perform manual gating from a FlowJo gate

perform manual gating by: reading a flowjo workspace file using the flowWorkspace library, and apply one of the gates to the inputflowFrame. Note that not only the selected gate, but also all its hierarchy of gates, will be applied to the input dataset.

applyFlowJoGate(
  ff,
  gateName,
  wspSource = c("import", "pData"),
  wspFile = NULL,
  pDataVar = NULL,
  pDataPathMapping = NULL,
  ...
)

Arguments

ff

a flowCore::flowFrame

gateName

the name of the flowJo gate that will be applied

wspSource

if "import", uses wspFile to read the flowjo workspace (should be a .wsp file) if "pData", uses pDataVar and pDataPathMapping to link a specific phenotype data variable to map different flowjo workspace files

wspFile

a flowjo workspace

pDataVar

variable name (column of pheno data) used to map the flowjo workspace file

pDataPathMapping

a named list:

item names are possible values of pDataVar

item values are character() providing the wspFile for the corresponding pDataVar value

...

Extra arguments passed to getFlowJoLabels()

Value

a list with :

the first element ("matrix") is a matrix containing filtering results for each specified gate

the second element ("labels") is a vector which assigns one label to each cell. If no cell type correspond to a specific cell, than the keyword 'unlabeled' is assigned to this cell. If the cell belongs to several gates (meaning that the gates ar not disjoints), than this cell is assigned to the gate with the less matching cells.